# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Dorra Alaseel - Complete API v1 Collection
  description: Complete API v1 endpoints for Dorra Alaseel Car Rental System - All
    /api/v1/ routes
  version: 1.0.0
servers:
  - url: "{{base_url}}"
...
paths:
...
  /api/v1/vendor/car/list:
    get:
      tags:
        - Vendor - Cars Management
      summary: List My Cars
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{vendor_token}}
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/cars:
    get:
      tags:
        - Cars (Public)
      summary: List All Cars
      description: Get all available cars with filtering and sorting
      parameters:
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
        - name: sort
          in: query
          schema:
            type: string
          description: price_asc, price_desc, name_asc, name_desc
          example: price_asc
        - name: per_page
          in: query
          schema:
            type: integer
          example: "15"
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/cars/{carId}:
    get:
      tags:
        - Cars (Public)
      summary: Get Car Details
      description: Get details of a specific car
      parameters:
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
        - name: carId
          in: path
          schema:
            type: integer
          required: true
          example: "123"
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
components:
...
  securitySchemes: {}